// crates/clip-mash-app/src/api.ts
export interface Clip {
  /** @minimum 0 */
  indexWithinMarker: number
  /** @minimum 0 */
  indexWithinVideo: number
  markerId: number
  markerTitle: string
  /** Start and endpoint inside the video in seconds. */
  range: [number, number]
  source: VideoSource
  videoId: string
}
// ...
export type VideoSource = (typeof VideoSource)[keyof typeof VideoSource]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VideoSource = {
  Folder: "Folder",
  Download: "Download",
  Stash: "Stash",
} as const